pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "self_improving_agents"
version = "0.1.0"
description = "A Python system for LLM prompt optimization"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "arize>=7.36.1",
    "arize-otel>=0.8.0",
    "arize-phoenix-evals>=0.20.4",
    "pydantic>=2.11.2",
    "wrapt>=1.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "openai>=1.70.0",
    "openinference-instrumentation-openai>=0.1.23",
    "python-dotenv>=1.1.0",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I"]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.isort]
known-first-party = ["self_improving_agents"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[dependency-groups]
dev = [

]
